+++ src/drmaster.cpp
@@ -1,4 +1,4 @@
-// $Id: drmaster.cpp 153 2019-09-01 14:27:02Z  $
+// $Id: drmaster.cpp 154 2020-03-05 15:36:24Z  $
 
 // Released under the GNU Public License, version 2
 //   see: https://www.gnu.org/licenses/gpl-2.0.html
@@ -38,7 +38,7 @@ using namespace std;
 
     <i>posn</i> is updated to point at the start of the next call
 */
-const string master_dta::_get_call(const string& contents, size_t& posn) const
+const string master_dta::_get_call(const string& contents, uint32_t& posn) const
 { string rv;
 
   while ((posn < contents.length()) and static_cast<int>(contents[posn]) != 0)
